Add Jest testing environment and tests for deque implementation and solved new task
This commit introduces the Jest testing framework to the project, as well as initial tests for the deque data structure implementation. @types/jest, jest, and ts-jest modules are added to the devDependencies. A new jest.config.js file is created to configure Jest with TypeScript. Tests in the new deque.test.ts file validate the core behavior of the deque data structure. New methods and properties in deque.ts file serve the proper functioning of the deque data structure.